+++ memory_free.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   memory_free.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: miki <miki@student.42.fr>                  +#+  +:+       +#+        */
+/*   By: mrosario <mrosario@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/05 18:23:53 by mrosario          #+#    #+#             */
-/*   Updated: 2021/03/19 03:24:51 by miki             ###   ########.fr       */
+/*   Updated: 2021/03/20 18:32:55 by mrosario         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -65,6 +65,8 @@ void	freeme(t_micli *micli)
 		micli->cmd_result_str = ft_del(micli->cmd_result_str);
 	if (micli->buffer)
 		micli->buffer = ft_del(micli->buffer);
+	if (micli->cmdhist.hist)
+		ft_free_split(micli->cmdhist.hist);
 	clear_cmdline(micli);
 	if (micli->tokdata.path_array)
 		micli->tokdata.path_array = ft_free_split(micli->tokdata.path_array);