+++ minishell.h
@@ -6,7 +6,7 @@
 /*   By: mrosario <mrosario@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/26 10:26:59 by mrosario          #+#    #+#             */
-/*   Updated: 2021/03/19 19:02:41 by mrosario         ###   ########.fr       */
+/*   Updated: 2021/03/20 19:08:10 by mrosario         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,9 +22,11 @@
 # include <dirent.h>
 # include <sys/wait.h>
 # include <signal.h>
+# include <termios.h>
 # include "libft.h"
 
-# define READLINE_BUFSIZE 1024
+# define READLINE_BUFSIZE 1
+# define CMDHIST_BUF 1
 # define BUILTINS "exit,pwd,export,env,echo,cd,unset"
 # define DQUOTE_ESC_CHARS "\"$\\"
 # define DEL 127
@@ -32,7 +34,7 @@
 # define NUL ""
 # define SYN_ERROR "micli: syntax error near unexpected token"
 
-typedef struct	s_tokendata
+typedef struct s_tokendata
 {
 	size_t			toksize;
 	size_t			args;
@@ -40,18 +42,18 @@ typedef struct	s_tokendata
 	char			*tok_start;
 	char			*tok_end;
 	char			*var_flag;
-	unsigned char	quote_flag:2;
-	unsigned char	redirect_flag:2;
-	unsigned char	escape_flag:1;
-	unsigned char	cmd_flag:1;
+	unsigned char	quote_flag : 2;
+	unsigned char	redirect_flag : 2;
+	unsigned char	escape_flag : 1;
+	unsigned char	cmd_flag : 1;
 }				t_tokendata;
 
-typedef struct	s_token
+typedef struct s_token
 {
 	t_list	*var_lst;
 }				t_token;
 
-typedef struct	s_cmdline
+typedef struct s_cmdline
 {
 	char			*cmd;
 	t_list			*arguments;
@@ -62,16 +64,24 @@ typedef struct	s_cmdline
 	char			**micli_argv;
 	int				fd_redir_in;
 	int				fd_redir_out;
-	unsigned char	redir_in_flag:1;
-	unsigned char	redir_out_flag:2;
+	unsigned char	redir_in_flag : 1;
+	unsigned char	redir_out_flag : 2;
 }				t_cmdline;
 
-typedef struct	s_builtins
+typedef struct s_cmdhist
+{
+	char			**hist;
+	size_t			cmdhist_buf;
+	size_t			ptrs_in_hist;
+	size_t			index;
+}				t_cmdhist;
+
+typedef struct s_builtins
 {
 	int	argflag;
 }				t_builtins;
 
-typedef struct	s_pipes
+typedef struct s_pipes
 {
 	int				*array;
 	size_t			array_size;
@@ -79,7 +89,7 @@ typedef struct	s_pipes
 	size_t			cmd_index;
 }				t_pipes;
 
-typedef struct	s_normis_fault
+typedef struct s_normis_fault
 {
 	size_t	pipe_max;
 	mode_t	perms;
@@ -88,23 +98,44 @@ typedef struct	s_normis_fault
 	int		f_re;
 }				t_normis_fault;
 
-typedef struct	s_micli
+typedef struct s_micli
 {
-	char			(*micli_loop)(struct s_micli*);
+	t_cmdhist		cmdhist;
 	t_tokendata		tokdata;
 	t_cmdline		cmdline;
 	t_token			token;
 	t_builtins		builtins;
 	t_pipes			pipes;
 	t_normis_fault	tonti;
+	struct termios	orig_term;
+	struct termios	raw_term;
 	int				syserror;
 	int				cmd_result;
 	char			*cmd_result_str;
 	char			**envp;
 	char			*buffer;
-	unsigned char	quote_flag:1;
-	unsigned char	pipe_flag:2;
+	unsigned char	quote_flag : 1;
+	unsigned char	pipe_flag : 2;
 }				t_micli;
+
+/*
+** Main Loop
+*/
+
+char			micli_loop(t_micli *micli);
+
+/*
+** Termcaps
+*/
+
+void			enable_raw_mode(struct termios *raw_term, \
+				struct termios *orig_term);
+
+/*
+** Command History
+*/
+
+void			cmdhist_ptr_array_alloc(t_micli *micli, t_cmdhist *cmdhist);
 
 /*
 ** Command Execution
@@ -207,7 +238,7 @@ char			*clean_ft_strdup(char const *str, t_micli *micli);
 char			*clean_ft_strjoin(char const *s1, char const *s2, \
 				t_micli *micli);
 char			**clean_ft_split(const char *s, char c, t_micli *micli);
-void			*ft_realloc(void *ptr, size_t size, t_micli *micli);
+void			*ft_realloc(void *ptr, size_t new_size, size_t old_size, t_micli *micli);
 void			*clean_calloc(size_t count, size_t size, t_micli *micli);
 
 /*